refactor: extract job catalog normalization
